Start revision F layout
The previous five revisions are not satisfactory. They have a number of
issues:

- Poor RF performance due to improper antenna placement and matching
- GPIO16 and GPIO17 are unusable, inherent problem with Pico D4
- Instability. Might be related to excessive power consumption due to
  improper antenna matching
- Changes in availability of LCSC parts

Overall those problems have prompted me to design a revision F of this
board. Most notably the ESP32 Pico D4 will be swapped for an ESP32-D0WD.
This will ensure the board actually behaves like a ESP32 DevKitC and
make it a true drop-in replacement. To ensure power and signal integrity
the new board will be a four layer one. Overall the board outline will
change slightly to allocate a suitable location for the antenna. With that
the height of the board will be closer to that of a DevKitC V1. Judging
from online listings the DevKitC V1 pinout still seems to be the most
popular one. Thus I will stick to that.
Also availability of cheap ESP32 development boards with USB Type C still
seems to be just as bad as back in 2020 when I designed the initial
revision of this board. Thus I think this new revision is a worthwhile
project.